Utilizing commands and removing '.disconnect'
