+++ src/psjReplServer.ts
@@ -2,17 +2,16 @@ import repl, { REPLServer } from 'repl';
 import { Context } from 'vm';
 import { CommandError, PicoSdkJsEngineConnection } from './PicoSdkJsEngineConnection';
 import { LogLevel, LogMessage, logger } from './psjLogger';
-import { connectToPico } from './commands/connectCommand';
-import { disconnectFromPico } from './commands/disconnectCommand';
-import { lsCommand } from './commands/lsCommand';
-import { statsCommand } from './commands/statsCommand';
-import { writeCommand } from './commands/writeCommand';
-import { readCommand } from './commands/readCommand';
-import { deleteCommand } from './commands/deleteCommand';
-import { formatCommand } from './commands/formatCommand';
-import { restartCommand } from './commands/restartCommand';
-import { killCommand } from './commands/killCommand';
-import { runCommand } from './commands/runCommand';
+import { connectToPico } from './replCommands/connectCommand';
+import { lsCommand } from './replCommands/lsCommand';
+import { statsCommand } from './replCommands/statsCommand';
+import { writeCommand } from './replCommands/writeCommand';
+import { readCommand } from './replCommands/readCommand';
+import { deleteCommand } from './replCommands/deleteCommand';
+import { formatCommand } from './replCommands/formatCommand';
+import { restartCommand } from './replCommands/restartCommand';
+import { killCommand } from './replCommands/killCommand';
+import { runCommand } from './replCommands/runCommand';
 
 export class PsjReplServer {
     private connection: PicoSdkJsEngineConnection | null = null;
@@ -39,7 +38,9 @@ export class PsjReplServer {
                 this.logFn(m);
             };
             this.connection.onClose = () => {
+                logger.logMsg(LogLevel.Error, 'Connection to Pico-SDK-JS engine lost.');
                 this.setConnection(null);
+                this.close();
             };
         }
     }
@@ -84,11 +85,6 @@ export class PsjReplServer {
         this.server.defineCommand('connect', {
             help: 'Connect to a Pico running Pico-Sdk-JS',
             action: (text: string) => this.wrapCommand(() => connectToPico(this, text))
-        });
-
-        this.server.defineCommand('disconnect', {
-            help: 'Disconnect a Pico running Pico-Sdk-JS',
-            action: () => this.wrapCommand(() => disconnectFromPico(this))
         });
 
         this.server.defineCommand('stats', {
