+++ src/replCommands/deleteCommand.ts
